* main: (29 commits)
  ftp --> resources
  Update release notes
  Convert the ENABLE_XXX options to NETCDF_ENABLE_XXX options
  Modify CMakeLists.txt to honor CMAKE_INSTALL_MANDIR in support of https://github.com/Unidata/netcdf-c/issues/2920.
  Modify ncdump to print char-valued variables as utf8.
  CI: Create an MSYS2/MinGW CMake run.
  CI: Test on MSYS2/UCRT64 environment
  CMake: Fix running tests on MinGW
  removal of ftp and contrib site
  removal of ftp site
  removal of ftp and netcdf contrib site
  removal of ftp and netcdf contrib site
  comment
  ftp --> resources for sample data location
  ftp--> resources for sample data location
  CMake: Enable plugins on MinGW
  Reintroduce targets to allow for cmake-based netCDF-Fortran to find and linka gainst netCDF-C. Stop-gap measure until we can modify netCDF-Fortran to use a more modern approach.
  revert
  Fix conversion warning in flag set/check/clear macros
  ckp
  ...

+++ libdap4/d4curlfunctions.c
@@ -363,8 +363,8 @@ NCD4_get_rcproperties(NCD4INFO* state)
 	    unsigned long interval=0;
 	    if(sscanf(option,"%lu/%lu",&idle,&interval) != 2)
 	        fprintf(stderr,"Illegal KEEPALIVE VALUE: %s\n",option);
-	    state->curl->keepalive.idle = idle;
-	    state->curl->keepalive.interval = interval;
+	    state->curl->keepalive.idle = (long)idle;
+	    state->curl->keepalive.interval = (long)interval;
 	    state->curl->keepalive.active = 1;
 	}
     }
